+++ src/main/java/de/samply/exporter/ExporterService.java
@@ -52,6 +52,7 @@ public class ExporterService {
     private final NotificationService notificationService;
     private final Set<String> exportTemplates;
     private final Set<String> datashieldTemplates;
+    private final Set<String> researcherEnvironmentTemplates;
     private final String focusProjectManagerId;
     private final String focusApiKey;
 
@@ -67,6 +68,7 @@ public class ExporterService {
             @Value(ProjectManagerConst.FOCUS_URL_SV) String focusUrl,
             @Value(ProjectManagerConst.EXPORT_TEMPLATES_SV) Set<String> exportTemplates,
             @Value(ProjectManagerConst.DATASHIELD_TEMPLATES_SV) Set<String> datashieldTemplates,
+            @Value(ProjectManagerConst.RESEARCHER_ENVIRONMENT_TEMPLATES_SV) Set<String> researcherEnvironmentTemplates,
             @Value(ProjectManagerConst.FOCUS_TTL_SV) String focusWaitTime,
             @Value(ProjectManagerConst.FOCUS_FAILURE_STRATEGY_MAX_TRIES_SV) String focusWaitCount,
             @Value(ProjectManagerConst.MAX_TIME_TO_WAIT_FOCUS_TASK_IN_MINUTES_SV) int maxTimeToWaitFocusTaskInMinutes,
@@ -84,6 +86,7 @@ public class ExporterService {
         this.focusWaitTime = focusWaitTime;
         this.focusWaitCount = focusWaitCount;
         this.maxTimeToWaitFocusTaskInMinutes = maxTimeToWaitFocusTaskInMinutes;
+        this.researcherEnvironmentTemplates = researcherEnvironmentTemplates;
         this.webClient = webClientFactory.createWebClient(focusUrl);
         this.focusApiKey = focusApiKey;
         this.projectBridgeheadRepository = projectBridgeheadRepository;
@@ -231,6 +234,7 @@ public class ExporterService {
         return switch (projectType) {
             case EXPORT -> exportTemplates;
             case DATASHIELD -> datashieldTemplates;
+            case RESEARCH_ENVIRONMENT -> researcherEnvironmentTemplates;
         };
     }